Add files via upload
It is the real time customer billing console application. It is the sort of mini project developed specially to generate customer billing for medical stores, cafes, super markets, shops etc. The entire program is written in C Programming Language and the program is compiled in VS Code using gcc compiler.